+++ test/integration/depth_camera_plugin.cc
@@ -249,8 +249,8 @@ void DepthCameraSensorTest::ImagesWithBuiltinSDF(
   // wait for a few depth camera frames
   mgr.RunOnce(ignition::common::Time::Zero, true);
 
-  int midWidth = depthSensor->ImageWidth() * 0.5;
-  int midHeight = depthSensor->ImageHeight() * 0.5;
+  int midWidth = static_cast<int>(depthSensor->ImageWidth() * 0.5);
+  int midHeight = static_cast<int>(depthSensor->ImageHeight() * 0.5);
   int mid = midHeight * depthSensor->ImageWidth() + midWidth -1;
   double expectedRangeAtMidPoint = boxPosition.X() - unitBoxSize * 0.5;
